requirements.txt
openai
requests
fpdf
pillow
firebase-admin
boto3
